refactor: renaming comments component file
